Refactor phpstan errors.

+++ app/Console/Commands/System/ForceDecimalSize.php
@@ -506,8 +506,8 @@ class ForceDecimalSize extends Command
                 continue;
             }
             // fix $field by rounding it down correctly.
-            $pow     = pow(10, (int)$currency->decimal_places);
-            $correct = bcdiv((string)round($value * $pow), (string)$pow, 12);
+            $pow     = (float)pow(10, (int)$currency->decimal_places);
+            $correct = bcdiv((string)round((float)$value * $pow), (string)$pow, 12);
             $this->friendlyWarning(sprintf('Transaction #%d has amount with value "%s", this has been corrected to "%s".', $item->id, $value, $correct));
             Transaction::find($item->id)->update(['amount' => $correct]);
         }
@@ -533,8 +533,8 @@ class ForceDecimalSize extends Command
                 continue;
             }
             // fix $field by rounding it down correctly.
-            $pow     = pow(10, (int)$currency->decimal_places);
-            $correct = bcdiv((string)round($value * $pow), (string)$pow, 12);
+            $pow     = (float)pow(10, (int)$currency->decimal_places);
+            $correct = bcdiv((string)round((float)$value * $pow), (string)$pow, 12);
             $this->friendlyWarning(
                 sprintf('Transaction #%d has foreign amount with value "%s", this has been corrected to "%s".', $item->id, $value, $correct)
             );
